Refactor and divide classes into packages Add TaskList class, Storage class
TODO: Add Ui and Parser class
