+++ data/output.txt
@@ -4,3 +4,4 @@ E | 0 | project meeting | Mon 2pm | 4pm
 D | 0 | return book | Oct 15 2019
 D | 0 | return book | 2 of December 2019, 6PM
 E | 0 | testEvent return book | Oct 15 2019 | 2 of December 2019, 6PM
+T | 0 | after refactor packages test
