+++ src/main/java/UI.java
@@ -1,0 +1,3 @@
+public class UI {
+
+}
