+++ src/main/java/constants/ExceptionErrorMessages.java
@@ -1,0 +1,8 @@
+package constants;
+
+public class ExceptionErrorMessages {
+    public static final String EMPTY_DESCRIPTION_ERROR_MESSAGE = "The description cannot be empty??";
+    public static final String NUMBER_FORMAT_ERROR_MESSAGE = "Invalid task number format! Please enter a valid number.";
+    public static final String NO_SUCH_TASK_NUMBER_ERROR_MESSAGE = "We do not have this task number!!";
+    public static final String FAILED_WRITE_TO_FILE_ERROR_MESSAGE = "Failed to write to file";
+}
