+++ src/main/java/constants/FilePaths.java
@@ -1,0 +1,6 @@
+package constants;
+
+public class FilePaths {
+    public static final String RELATIVE_DATA_DIRECTORY_PATH = "./data";
+    public static final String RELATIVE_OUTPUT_TXT_FILE_PATH = "./data/output.txt";
+}
