+++ src/main/java/constants/Formats.java
@@ -1,0 +1,6 @@
+package constants;
+
+public class Formats {
+    public static final String EVENT_FORMAT = "event [task] /from [date] /to [date]";
+    public static final String DEADLINE_FORMAT = "deadline [task] /by [date]";
+}
