+++ src/main/java/constants/Information.java
@@ -1,0 +1,6 @@
+package constants;
+
+public class Information {
+    public static final String solidLineBreak = "____________________________________________________________";
+    public static final String CHATBOT_NAME = "ByteBuddy";
+}
