+++ src/main/java/constants/Messages.java
@@ -1,0 +1,8 @@
+package constants;
+
+import static constants.Information.CHATBOT_NAME;
+
+public class Messages {
+    public static final String START_MESSAGE = "Hello! I'm " + CHATBOT_NAME + "\n" + "\t What can I do for you?";
+    public static final String BYE_MESSAGE = "Sad to see you leave :(";
+}
