+++ src/main/java/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package tasks;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
